Starting to split between source files
