+++ inavjaga.cpp
@@ -973,20 +973,6 @@ ANSI::Settings Chest::chestStyle = {
     ANSI::Attribute::REVERSE
 };
 
-Portal::Portal(sista::Coordinates coordinates) : Entity('&', coordinates, portalStyle, Type::PORTAL) {
-    Portal::portals.push_back(this);
-}
-void Portal::remove() {
-    Portal::portals.erase(std::find(Portal::portals.begin(), Portal::portals.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-ANSI::Settings Portal::portalStyle = {
-    RGB_ROCKS_FOREGROUND,
-    RGB_ROCKS_BACKGROUND,
-    ANSI::Attribute::FAINT
-};
-
 Mine::Mine(sista::Coordinates coordinates) : Entity('*', coordinates, mineStyle, Type::MINE), triggered(false) {
     Mine::mines.push_back(this);
 }
